# [Do not merge] ASP.NET Core tracing perf summary