--- a/PerfTestAspNetCore/InstrumentationOptions.cs
+++ b/PerfTestAspNetCore/InstrumentationOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace PerfTestAspNetCore
+{
+    public class InstrumentationOptions
+    {
+        public bool EnableDiagnosticSource { get; set; }
+
+        public bool EnableOTel { get; set; }
+
+        public bool EnableMiddleware { get; set; }
+    }
+}