--- a/PerfTestAspNetCore/appsettings.Development.json
+++ b/PerfTestAspNetCore/appsettings.Development.json
@@ -1,0 +1,8 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  }
+}