--- a/PerfTestAspNetCore/appsettings.json
+++ b/PerfTestAspNetCore/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "InstrumentationOptions": {
+    "EnableDiagnosticSource": false,
+    "EnableOTel": false,
+    "EnableMiddleware": false
+  },
+  "AllowedHosts": "*"
+}