Towards #4214

## Changes
This PR is summarizing the issue raised in #4214. The numbers shared in the original issue does not show 1:1 comparison of instrumentation done via middleware versus diagnostic source callbacks. The test app shown in this PR attempts to do 1:1 comparison to find the root cause of overhead.

The tests are divided into 4 categories and performed using https://github.com/codesenberg/bombardier:

`bombardier-windows-amd64.exe -c 1 -n 1000000 http://localhost:5152/weatherforecast`


**No instrumentation**

```
Statistics        Avg      Stdev        Max
  Reqs/sec      5603.76     314.83    6520.68
  Latency      175.29us   102.62us    99.70ms
  HTTP codes:
    1xx - 0, 2xx - 1000000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.33MB/s
```

**Instrumentation done via middleware**
```
Reqs/sec      5324.58     298.88    6118.92
  Latency      183.84us   115.64us   113.10ms
  HTTP codes:
    1xx - 0, 2xx - 1000000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.17MB/s
```

**Enabling Diagnostic Source callbacks without activity enrichment (This is to show the overhead caused by just subscribing to events)**
```
Reqs/sec      5334.77     270.81    6340.36
  Latency      183.59us   100.17us    97.32ms
  HTTP codes:
    1xx - 0, 2xx - 1000000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.17MB/s
```

**Instrumentation done via instrumentation library**
```
Statistics        Avg      Stdev        Max
  Reqs/sec      5093.24     254.27    6031.88
  Latency      194.53us   100.95us    97.94ms
  HTTP codes:
    1xx - 0, 2xx - 1000000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.03MB/s
```

In **summary**:
Instrumentation via middleware causes **~5%** overhead
Instrumentation via instrumentation library causes **~10%** overhead with ~**5%** caused by DiagnosticSource events subscription.

## Merge requirement checklist

* [ ] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (nullable enabled, static analysis, etc.)
* [ ] Unit tests added/updated
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial changes
* [ ] Changes in public API reviewed (if applicable)